[32/64-bit] CLI builds in 32/64-bit
The CLI project now builds with a simple:

make 32-bit
make 64-bit

Quick check of the cli.list output suggests things look vaguely sensible, so this should allow
the checking of the 32-bit build initially against the full 32-bit kernel environment, which should
then provide SOME degree of confidence that once the 64-bit environment fires up completely, it should
be able to run the 64-bit CLI variant without the CLI code itself being of any obvious issue.

+++ c/common/main.c
@@ -16,9 +16,9 @@
 
 
 
-#include "stdlib/include/stdio.h"
-#include "stdlib/include/kernel.h"
-#include "stdlib/include/swis.h"
+#include <stdint.h>
+#include "stdio.h"
+#include "swi.h"
 
 
 
@@ -69,7 +69,7 @@ void execute_command( void ) {
     input[i]   = 0x00;
 
     struct _kernel_regs in,out;
-    in.r[0] = (unsigned int) input;
+    in.r[0] = (uintptr_t) input;
     _kernel_swi( OS_ProcessBegin, &in, &out );
 	sprintf( "(and now CLI can continue)\n" );
